refactor(testing): fix_damage/fix_randomを公開APIから除外
対戦オブジェクトの内部属性を直接差し替えるモンキーパッチであり、jpoke.testingが
保証すべきAPI安定性の対象にそぐわないため、リポジトリ内テスト専用としてtests/test_utils.pyへ移設した。
